requirements.txt
httpx>=0.24.0
mcp-server>=0.1.0
python-dotenv>=1.0.0
google-api-python-client>=2.90.0
oauth2client>=4.1.3
